+++ sed3/__init__.py
@@ -1,2 +1,3 @@
 __all__ = ["sed3"]
 from .sed3 import *
+__version__ = "1.4.3"