+++ changelog.md
@@ -4,6 +4,9 @@
 
 # 1.4.
 
+* [Added] `imin` and `imax` to control intensity window slicers
+* [Fixed] Window slicer update on set_window function
+* [Added] Keyword `contours` with same meaning as `contour` 
 * [Changed] PyQt5 compatibility
 * [Changed] Using `pytest` instead of `nose`